+++ setup.py
@@ -18,7 +18,6 @@ author='Bartosz Bosak',
 install_requires=[
     'pytest',
     'pytest-pep8',
-    'scipy==1.2.1',
     'qcgPilotManager @ git+https://github.com/vecma-project/QCG-PilotJob.git@v0.6#egg=qcgPilotManager',
     'easyvvuq @ git+https://github.com/UCL-CCS/EasyVVUQ.git@v0.5#egg=easyvvuq'
 ],
